ProgController
ProgramaController Modificado

+++ src/main/java/com/example/demo/Controller/ProgramaController.java
@@ -1,20 +1,21 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
+
 package com.example.demo.Controller;
 
 import com.example.demo.Entity.Programa;
 import com.example.demo.Service.ProgramaService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
+import java.util.HashMap;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -23,6 +24,7 @@ import org.springframework.web.bind.annotation.RestController;
  *
  * @author franc
  */
+@CrossOrigin(origins = {"http://localhost:4200"})
 @RestController
 @RequestMapping("/programa")
 @Api(value = "Microservicio de programa", description = "Microservicio de programa")
@@ -30,28 +32,69 @@ public class ProgramaController {
      @Autowired
     ProgramaService programaService;
 
-    @ApiOperation(value = "Lista de programas")
+ @ApiOperation(value = "Lista de programas")
     @GetMapping
-    public List<Programa> findAll() {
-        return (List<Programa>) programaService.findAll();
+    public ResponseEntity<?> findAll() {
+        HashMap<String, Object> result = new HashMap<>();
+        result.put("success", true);
+        result.put("message", "Lista de programas");
+        result.put("data", programaService.findAll());
+        return new ResponseEntity<>(result, HttpStatus.OK);
     }
 
-    @ApiOperation(value = "Lista un libro")
-    @GetMapping(value = "/{id}")
+    @ApiOperation(value = "Obtiene datos de una programa")
+    @GetMapping("/{id}")
     public ResponseEntity<Programa> findById(@PathVariable Long id) {
         Programa programa = programaService.findById(id);
         return ResponseEntity.ok(programa);
     }
 
-    @ApiOperation(value = "registrar un libro")
+    @ApiOperation(value = "Crea una programa")
     @PostMapping
-    public Programa save(@RequestBody Programa programa) {
-        return programaService.save(programa);
+    public ResponseEntity<?> save(@RequestBody Programa programa) {
+        HashMap<String, Object> result = new HashMap<>();
+        result.put("success", true);
+        result.put("message", "Programa registrado correctamente");
+        result.put("data", programaService.save(programa));
+        return new ResponseEntity<>(result, HttpStatus.OK);
     }
 
-    @ApiOperation(value = "Elimina una libro")
-    @DeleteMapping(value = "/{id}")
-    public void deleteById(@PathVariable Long id) {
-        programaService.deleteById(id);
+    @ApiOperation(value = "Modifica una programa")
+    @PutMapping("/{id}")
+    public ResponseEntity<?> update(@PathVariable(value = "id") Long id, @RequestBody Programa programa) {
+        HashMap<String, Object> result = new HashMap<>();
+        Programa data = programaService.findById(id);
+        if (data == null) {
+            result.put("success", false);
+            result.put("message", "No existe programa con Id: " + id);
+            return new ResponseEntity<>(result, HttpStatus.NOT_FOUND);
+        }
+        try {
+            programa.setProgId(id);
+            programaService.save(programa);
+            result.put("success", true);
+            result.put("message", "Datos actualizados correctamente.");
+            result.put("data", programa);
+            return new ResponseEntity<>(result, HttpStatus.OK);
+        } catch (Exception ex) {
+            return new ResponseEntity<>(new Exception(ex.getMessage()), HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+
+    @ApiOperation(value = "Elimina una programa")
+    @DeleteMapping("/{id}")
+    public ResponseEntity<?> deleteById(@PathVariable Long id) {
+        HashMap<String, Object> result = new HashMap<>();
+        Programa data = programaService.findById(id);
+        if (data == null) {
+            result.put("success", false);
+            result.put("message", "No existe programa con id: " + id);
+            return new ResponseEntity<>(result, HttpStatus.NOT_FOUND);
+        } else {
+            programaService.deleteById(id);
+            result.put("success", true);
+            result.put("message", "Registro eliminado correctamente.");
+            return new ResponseEntity<>(result, HttpStatus.OK);
+        }
     }
 }
